Move rsspp::RssParser into a separate file
