+++ rss/rssparser.cpp
@@ -1,4 +1,4 @@
-#include "rssppinternal.h"
+#include "rssparser.h"
 
 #include <cstring>
 #include <libxml/tree.h>
